[SYSTEMML-2184] Apply basic wdivmm conservatively (sparse/dense inputs)
This patch makes a compiler improvement to apply the fused basic wdivmm
for the pattern (X * U%*%t(V)) more conservatively only for when (1) X
is sparse and U/V are unknown or dense, or (2) when U and V known to be
dense, which are the primary applications for this sparsity exploiting
fused operator. At the same time, this helps to avoid performance issues
for scenarios where U and V are actually ultra-sparse and wdivmm would
hurt performance.

For example, on perftest stratstats 10K x 10K, this patch improved
performance from 128s to 29s end-to-end.

+++ src/test/java/org/apache/sysml/test/integration/AutomatedTestBase.java
@@ -504,14 +504,12 @@ public abstract class AutomatedTestBase
 		return matrix;
 	}
 
-	protected double[][] writeInputMatrixWithMTD(String name, double[][] matrix, boolean bIncludeR)
-	{
+	protected double[][] writeInputMatrixWithMTD(String name, double[][] matrix, boolean bIncludeR) {
 		MatrixCharacteristics mc = new MatrixCharacteristics(matrix.length, matrix[0].length, OptimizerUtils.DEFAULT_BLOCKSIZE, OptimizerUtils.DEFAULT_BLOCKSIZE, -1);
 		return writeInputMatrixWithMTD(name, matrix, bIncludeR, mc);
 	}
 
-	protected double[][] writeInputMatrixWithMTD(String name, double[][] matrix, int nnz, boolean bIncludeR)
-	{
+	protected double[][] writeInputMatrixWithMTD(String name, double[][] matrix, long nnz, boolean bIncludeR) {
 		MatrixCharacteristics mc = new MatrixCharacteristics(matrix.length, matrix[0].length, OptimizerUtils.DEFAULT_BLOCKSIZE, OptimizerUtils.DEFAULT_BLOCKSIZE, nnz);
 		return writeInputMatrixWithMTD(name, matrix, bIncludeR, mc);
 	}
